agentTemplate/pyproject.toml
[project]
name = "a2a-agent-template"
version = "0.1.0"
description = "Pluggable A2A + LangGraph Agent Template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core A2A and LangGraph dependencies
    "a2a-sdk>=0.2.6,<0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.1.0",
    "langgraph>=0.3.18",
    
    # MCP integration
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.0.0",
    
    # HTTP and API support
    "httpx>=0.28.1",
    "aiohttp>=3.8.0",
    
    # Server and utilities
    "uvicorn>=0.34.2",
    "click>=8.1.8",
    "python-dotenv>=1.1.0",
    "pydantic>=2.10.6",
    "psutil>=5.9.0",
    
    # Optional: Common tool libraries
    "requests>=2.31.0",
    "jsonschema>=4.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "config", "plugins"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto" 